backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.3",
    "fastapi>=0.135.1",
    "fpdf2>=2.8.7",
    "google-genai>=1.66.0",
    "langchain-google-genai>=4.2.1",
    "langgraph>=1.0.10",
    "newsapi-python>=0.2.7",
    "pydantic>=2.12.5",
    "pymupdf>=1.27.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "uvicorn>=0.41.0",
    "yfinance>=1.2.0",
]
